fature get id productplatform get id logoproduct deleted and merged with vendorproduct

+++ src/main/java/com/winningstation/services/VendorProductService.java
@@ -1,80 +1,103 @@
 package com.winningstation.services;
 
-import com.winningstation.entity.PlatformProduct;
 import com.winningstation.entity.VendorProduct;
-import com.winningstation.repository.PlatformProductRepository;
 import com.winningstation.repository.VendorProductRepository;
 import com.winningstation.services.interfaces.IVendorProductService;
 import jakarta.transaction.Transactional;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
 
 /**
- * Servicio que permite realizar operaciones sobre la entidad VendorProduct.
+ * Servicio que permite realizar operaciones sobre los logos de los productos.
  *
  * @author David Portillo Hoyos
  */
 @Service
 @Transactional
 public class VendorProductService implements IVendorProductService {
-  /** Logger */
+
+  /** Logger de la clase. */
   private static final Logger LOGGER = LoggerFactory.getLogger(VendorProductService.class);
 
-  /** Repositorio de VendorProduct */
+  /** Repositorio de la entidad LogoProduct. */
   private final VendorProductRepository vendorProductRepository;
 
+  private final FileStorageService fileStorageService;
+
   /**
-   * Constructor
+   * Constructor de la clase.
    *
-   * @param vendorProductRepository Repositorio de VendorProduct
+   * @param vendorProductRepository Repositorio de la entidad LogoProduct.
    */
-  public VendorProductService(VendorProductRepository vendorProductRepository) {
+  public VendorProductService(
+          VendorProductRepository vendorProductRepository, FileStorageService fileStorageService) {
     this.vendorProductRepository = vendorProductRepository;
+    this.fileStorageService = fileStorageService;
   }
 
   @Override
-  public VendorProduct save(VendorProduct vendorProduct) {
-    LOGGER.info("Guardando VendorProduct {}", vendorProduct);
-    return vendorProductRepository.save(vendorProduct);
+  public VendorProduct save(VendorProduct logo, MultipartFile file) {
+    LOGGER.info("Guardando logo de producto: {}", logo);
+    String fileDownloadUri = fileStorageService.storeFileAndGenerateUri(file);
+    logo.setLogo(fileDownloadUri);
+    return vendorProductRepository.save(logo);
   }
 
   @Override
   public VendorProduct findById(Long id) {
-    LOGGER.info("Obteniendo VendorProduct por id {}", id);
-    return vendorProductRepository.findById(id).orElse(null);
+    if (vendorProductRepository.findById(id).isPresent()) {
+      return vendorProductRepository.findById(id).get();
+    } else {
+      LOGGER.error("No se encontró el logo de producto con id: {}", id);
+      throw new RuntimeException("No se encontró el logo de producto con id: " + id);
+    }
   }
 
   @Override
   public List<VendorProduct> findAll() {
-    LOGGER.info("Obteniendo todos los VendorProduct");
     return vendorProductRepository.findAll();
   }
 
   @Override
   public void deleteById(Long id) {
-    LOGGER.info("Eliminando VendorProduct por id {}", id);
-    vendorProductRepository.deleteById(id);
+    if (vendorProductRepository.existsById(id)) {
+      vendorProductRepository.deleteById(id);
+    } else {
+      LOGGER.error("No se encontró el logo de producto con id: {}", id);
+      throw new RuntimeException("No se encontró el logo de producto con id: " + id);
+    }
   }
 
   @Override
-  public String update(Long id, String request) {
-    LOGGER.info("Actualizando VendorProduct por id {}", id);
-    VendorProduct vendorProduct = vendorProductRepository.findById(id).orElse(null);
-    if (vendorProduct != null) {
-      vendorProduct.setName(request);
-      vendorProductRepository.save(vendorProduct);
-      return request;
+  public VendorProduct update(Long id, VendorProduct newVendorProduct, MultipartFile file) {
+    VendorProduct existingVendorProduct = vendorProductRepository.findById(id).orElse(null);
+    if (existingVendorProduct != null) {
+      LOGGER.info("Actualizando logo de producto con id: {}", id);
+      if (file != null) {
+        String fileDownloadUri = fileStorageService.replaceFileAndGenerateUri(file, existingVendorProduct.getLogo());
+        existingVendorProduct.setLogo(fileDownloadUri);
+      }
+      if (newVendorProduct.getAlt() != null) {
+        existingVendorProduct.setAlt(newVendorProduct.getAlt());
+      }
+
+      if (newVendorProduct.getName() != null) {
+        existingVendorProduct.setName(newVendorProduct.getName());
+      }
+
+      return vendorProductRepository.save(existingVendorProduct);
     } else {
-      throw new RuntimeException("Registro no encontrado");
+      LOGGER.error("No se encontró el logo de producto con id: {}", id);
+      throw new RuntimeException("No se encontró el logo de producto con id: " + id);
     }
   }
 
   @Override
   public List<VendorProduct> findByName(String name) {
-    LOGGER.info("Obteniendo VendorProduct por nombre {}", name);
-    return vendorProductRepository.findByNameContaining(name);
+    return vendorProductRepository.findByLogoContaining(name);
   }
 }
